refactor(renderer/api/update): move interfaces to co-located types/ folder
