+++ src/renderer/api/update/versionApi.ts
@@ -24,15 +24,9 @@
  * @author 鸡哥
  */
 
-/** 版本信息接口 */
-export interface VersionInfo {
-  appName: string;
-  version: string;
-  description: string;
-  downloadUrl: string;
-  id: number;
-  updatedAt: string;
-}
+import type { VersionInfo } from './types/VersionInfo';
+
+export type { VersionInfo };
 
 const IS_DEV_RENDERER = window.location.hostname === 'localhost' || window.location.hostname === '127.0.0.1';
 const VERSION_API_BASE = IS_DEV_RENDERER
